add command to change image paths in notebooks
needed this to work with slides.

during writing, point images to ../fig/blahblah
this script then changes all those links to ./fig/blahblah (single dot).
the idea is to copy the fig folder to the notebooks folder before starting
the presentation. A hacky way to make them work with this setup.